requirements.txt
snowflake-connector-python>=3.0.0
croniter>=1.3.0
